+++ lib/tasks/bundle/update.rb
@@ -1,0 +1,13 @@
+module Rsm
+  module Bundle
+    class Update < Rsm::Base
+
+      def update
+        inside "." do
+          run_ruby_binary "bundle update"
+        end
+      end
+
+    end
+  end
+end